+++ source/icu.net/BreakIterator.cs
@@ -2,11 +2,17 @@
 // This software is licensed under the MIT license (http://opensource.org/licenses/MIT)
 using System;
 using System.Collections.Generic;
-using System.Linq;
 
 namespace Icu
 {
-	public static class BreakIterator
+	/// <summary>
+	/// The BreakIterator implements methods for finding the location of
+	/// boundaries in text.
+	/// When using BreakIterator class, it will iterate over the boundaries
+	/// as described here: http://userguide.icu-project.org/boundaryanalysis
+	/// for all UBreakIteratorTypes (including UBreakIteratorType.Word).
+	/// </summary>
+	public abstract class BreakIterator : IDisposable
 	{
 		/// <summary>
 		/// The possible types of text boundaries.
@@ -37,13 +43,38 @@ namespace Icu
 			/// Upper bound for tags for uncategorized words.
 			/// </summary>
 			NONE_LIMIT     = 100,
+			/// <summary>
+			/// Tag value for words that appear to be numbers, lower limit.
+			/// </summary>
 			NUMBER         = 100,
+			/// <summary>
+			/// Tag value for words that appear to be numbers, upper limit.
+			/// </summary>
 			NUMBER_LIMIT   = 200,
+			/// <summary>
+			/// Tag value for words that contain letters, excluding hiragana,
+			/// katakana or ideographic characters, lower limit.
+			/// </summary>
 			LETTER         = 200,
+			/// <summary>
+			/// Tag value for words containing letters, upper limit.
+			/// </summary>
 			LETTER_LIMIT   = 300,
+			/// <summary>
+			/// Tag value for words containing kana characters, lower limit.
+			/// </summary>
 			KANA           = 300,
+			/// <summary>
+			/// Tag value for words containing kana characters, upper limit.
+			/// </summary>
 			KANA_LIMIT     = 400,
+			/// <summary>
+			/// Tag value for words containing ideographic characters, lower limit.
+			/// </summary>
 			IDEO           = 400,
+			/// <summary>
+			/// Tag value for words containing ideographic characters, upper limit.
+			/// </summary>
 			IDEO_LIMIT     = 500,
 		}
 
@@ -69,8 +100,175 @@ namespace Icu
 		public const int DONE = -1;
 
 		/// <summary>
-		/// Splits the specified text along the specified type of boundaries. Spaces and punctuations
-		/// are not returned.
+		/// Gets all of the boundaries for the given text.
+		/// </summary>
+		public abstract int[] Boundaries { get; }
+
+		/// <summary>
+		/// Gets the text being examined by this BreakIterator.
+		/// </summary>
+		public abstract string Text { get; }
+
+		/// <summary>
+		/// Determine the most recently-returned text boundary.
+		/// Returns <see cref="DONE"/> if there are no boundaries left to return.
+		/// </summary>
+		public abstract int Current { get; }
+
+		/// <summary>
+		/// Decrements the iterator and returns the previous boundary.
+		/// Returns DONE if the iterator moves past the first boundary.
+		/// </summary>
+		public abstract int MovePrevious();
+
+		/// <summary>
+		/// Increments the iterator and returns the next boundary.
+		/// Returns DONE if there are no boundaries left to return.
+		/// </summary>
+		public abstract int MoveNext();
+
+		/// <summary>
+		/// Sets the iterator to the first boundary and returns it.
+		/// Returns DONE if there was no text set.
+		/// </summary>
+		public abstract int MoveFirst();
+
+		/// <summary>
+		/// Sets the iterator to the last boundary and returns the offset into
+		/// the text.
+		/// Returns DONE if there was no text set.
+		/// </summary>
+		public abstract int MoveLast();
+
+		/// <summary>
+		/// Sets the iterator to refer to the first boundary position following
+		/// the specified position.
+		/// </summary>
+		/// <param name="offset">The position from which to begin searching for
+		/// a break position.</param>
+		/// <returns>The position of the first break after the current position.
+		/// The value returned is always greater than offset, or <see cref="BreakIterator.DONE"/>
+		/// </returns>
+		public abstract int MoveFollowing(int offset);
+
+		/// <summary>
+		/// Sets the iterator to refer to the last boundary position before the
+		/// specified position.
+		/// </summary>
+		/// <param name="offset">The position to begin searching for a break from.</param>
+		/// <returns>The position of the last boundary before the starting
+		/// position. The value returned is always smaller than the offset
+		/// or the value <see cref="BreakIterator.DONE"/></returns>
+		public abstract int MovePreceding(int offset);
+
+		/// <summary>
+		/// Returns true if the specified position is a boundary position and
+		/// false, otherwise. In addition, it leaves the iterator pointing to
+		/// the first boundary position at or after "offset".
+		/// </summary>
+		public abstract bool IsBoundary(int offset);
+
+		/// <summary>
+		/// Returns the status tag from the break rule that determined the
+		/// current position.  For break iterator types that do not support a
+		/// rule status, a default value of 0 is returned.
+		/// </summary>
+		/// <remarks>
+		/// For more information, see
+		/// http://userguide.icu-project.org/boundaryanalysis#TOC-Rule-Status-Values
+		/// </remarks>
+		public abstract int GetRuleStatus();
+
+		/// <summary>
+		/// Get the statuses from the break rules that determined the most
+		/// recently returned break position.
+		///
+		/// The values appear in the rule source within brackets, {123}, for
+		/// example.  The default status value for rules that do not explicitly
+		/// provide one is zero.
+		///
+		/// For word break iterators, the possible values are defined in enum
+		/// <see cref="BreakIterator.UWordBreak"/>.
+		/// </summary>
+		/// <remarks>
+		/// For more information, see
+		/// http://userguide.icu-project.org/boundaryanalysis#TOC-Rule-Status-Values
+		/// </remarks>
+		public abstract int[] GetRuleStatusVector();
+
+		/// <summary>
+		/// Sets the current text being examined to the given text.
+		/// Sets the current index back to the first element.
+		/// </summary>
+		/// <param name="text">New text.</param>
+		/// <exception cref="ArgumentNullException">Thrown when given text is
+		/// null. </exception>
+		public abstract void SetText(string text);
+
+		/// <summary>
+		/// Gets the locale for this BreakIterator.
+		/// </summary>
+		public abstract Locale Locale { get; }
+
+		/// <summary>
+		/// Creates a BreakIterator that splits on characters for the given locale.
+		/// </summary>
+		/// <param name="locale">The locale.</param>
+		public static BreakIterator CreateCharacterInstance(Locale locale)
+		{
+			return new RuleBasedBreakIterator(UBreakIteratorType.CHARACTER, locale);
+		}
+
+		/// <summary>
+		/// Creates a BreakIterator that splits on words for the given locale.
+		/// It iterates over boundaries as described in the "Word Boundary"
+		/// section http://userguide.icu-project.org/boundaryanalysis.
+		/// If you want to ignore spaces and punctuation, consider using:
+		/// <see cref="BreakIterator.Split(UBreakIteratorType, Locale, string)"/>
+		/// or <see cref="BreakIterator.GetWordBoundaries(Locale, string, bool)"/>,
+		/// </summary>
+		/// <param name="locale">The locale.</param>
+		public static BreakIterator CreateWordInstance(Locale locale)
+		{
+			return new RuleBasedBreakIterator(UBreakIteratorType.WORD, locale);
+		}
+
+		/// <summary>
+		/// Creates a BreakIterator that splits on lines for the given locale.
+		/// </summary>
+		/// <param name="locale">The locale.</param>
+		/// <param name="text">The initial text.</param>
+		public static BreakIterator CreateLineInstance(Locale locale)
+		{
+			return new RuleBasedBreakIterator(UBreakIteratorType.LINE, locale);
+		}
+
+		/// <summary>
+		/// Creates a BreakIterator that splits on sentences for the given locale.
+		/// </summary>
+		/// <param name="locale">The locale.</param>
+		/// <param name="text">The initial text.</param>
+		public static BreakIterator CreateSentenceInstance(Locale locale)
+		{
+			return new RuleBasedBreakIterator(UBreakIteratorType.SENTENCE, locale);
+		}
+
+		/// <summary>
+		/// Splits the specified text along the specified type of boundaries.
+		/// Spaces and punctuations are not returned.
+		/// </summary>
+		/// <param name="type">The type.</param>
+		/// <param name="locale">The locale.</param>
+		/// <param name="text">The text.</param>
+		/// <returns>The tokens.</returns>
+		public static IEnumerable<string> Split(UBreakIteratorType type, string locale, string text)
+		{
+			return Split(type, new Locale(locale), text);
+		}
+
+		/// <summary>
+		/// Splits the specified text along the specified type of boundaries.
+		/// Spaces and punctuations are not returned.
 		/// </summary>
 		/// <param name="type">The type.</param>
 		/// <param name="locale">The locale.</param>
@@ -78,15 +276,10 @@ namespace Icu
 		/// <returns>The tokens.</returns>
 		public static IEnumerable<string> Split(UBreakIteratorType type, Locale locale, string text)
 		{
-			return Split(type, locale.Id, text);
-		}
-
-		public static IEnumerable<string> Split(UBreakIteratorType type, string locale, string text)
-		{
 			if (string.IsNullOrEmpty(text))
 				yield break;
 
-			foreach (var boundary in GetBoundaries(type, locale, text, includeSpacesAndPunctuation: false))
+			foreach (var boundary in GetBoundaries(type, locale.Id, text, includeSpacesAndPunctuation: false))
 			{
 				yield return text.Substring(boundary.Start, boundary.End - boundary.Start);
 			}
@@ -95,17 +288,27 @@ namespace Icu
 		/// <summary>
 		/// Gets word boundaries for given text.
 		/// </summary>
-		/// <param name="includeSpacesAndPunctuation">ICU's UBreakIteratorType.WORD analysis considers
-		/// spaces and punctuation as boundaries for words. Set parameter to true if all boundaries
-		/// are desired; false otherwise.
+		/// <param name="includeSpacesAndPunctuation">
+		/// ICU's UBreakIteratorType.WORD analysis considers spaces and
+		/// punctuation as boundaries for words. Set parameter to true if all
+		/// boundaries are desired; false otherwise.
+		/// For more information: http://userguide.icu-project.org/boundaryanalysis#TOC-Count-the-words-in-a-document-C-only-:
+		/// </param>
+		public static IEnumerable<Boundary> GetWordBoundaries(string locale, string text, bool includeSpacesAndPunctuation)
+		{
+			return GetWordBoundaries(new Locale(locale), text, includeSpacesAndPunctuation);
+		}
+
+		/// <summary>
+		/// Gets word boundaries for given text.
+		/// </summary>
+		/// <param name="includeSpacesAndPunctuation">
+		/// ICU's UBreakIteratorType.WORD analysis considers spaces and
+		/// punctuation as boundaries for words. Set parameter to true if all
+		/// boundaries are desired; false otherwise.
 		/// For more information: http://userguide.icu-project.org/boundaryanalysis#TOC-Count-the-words-in-a-document-C-only-:
 		/// </param>
 		public static IEnumerable<Boundary> GetWordBoundaries(Locale locale, string text, bool includeSpacesAndPunctuation)
-		{
-			return GetWordBoundaries(locale.Id, text, includeSpacesAndPunctuation);
-		}
-
-		public static IEnumerable<Boundary> GetWordBoundaries(string locale, string text, bool includeSpacesAndPunctuation)
 		{
 			return GetBoundaries(UBreakIteratorType.WORD, locale, text, includeSpacesAndPunctuation);
 		}
@@ -116,40 +319,39 @@ namespace Icu
 		/// </summary>
 		public static IEnumerable<Boundary> GetBoundaries(UBreakIteratorType type, Locale locale, string text)
 		{
-			return GetBoundaries(type, locale.Id, text, false);
+			return GetBoundaries(type, locale, text, false);
 		}
 
-		private static IEnumerable<Boundary> GetBoundaries(UBreakIteratorType type, string locale, string text, bool includeSpacesAndPunctuation)
+		private static IEnumerable<Boundary> GetBoundaries(UBreakIteratorType type, Locale locale, string text, bool includeSpacesAndPunctuation)
 		{
-			if (string.IsNullOrEmpty(text))
-				yield break;
+			List<Boundary> boundaries = new List<Boundary>();
 
-			ErrorCode err;
-			IntPtr bi = NativeMethods.ubrk_open(type, locale, text, text.Length, out err);
-			if (err.IsFailure())
-				throw new Exception("BreakIterator.Split() failed with code " + err);
-
-			int cur = NativeMethods.ubrk_first(bi);
-
-			while (cur != DONE)
+			using (var breakIterator = new RuleBasedBreakIterator(type, locale))
 			{
-				int next = NativeMethods.ubrk_next(bi);
-				int status = NativeMethods.ubrk_getRuleStatus(bi);
-				
-				if (next == DONE)
-				{
-					break;
-				}
+				breakIterator.SetText(text);
 
-				if (includeSpacesAndPunctuation || AddToken(type, status))
-				{
-					yield return new Boundary(cur, next);
-				}
+				int current = breakIterator.Current;
 
-				cur = next;
+				while (current != DONE)
+				{
+					int next = breakIterator.MoveNext();
+					int status = breakIterator.GetRuleStatus();
+
+					if (next == DONE)
+					{
+						break;
+					}
+
+					if (includeSpacesAndPunctuation || AddToken(type, status))
+					{
+						boundaries.Add(new Boundary(current, next));
+					}
+
+					current = next;
+				}
 			}
 
-			NativeMethods.ubrk_close(bi);
+			return boundaries;
 		}
 
 		private static bool AddToken(UBreakIteratorType type, int status)
@@ -166,5 +368,16 @@ namespace Icu
 			}
 			return false;
 		}
+
+		/// <summary>
+		/// Dispose of managed/unmanaged resources.
+		/// </summary>
+		public void Dispose()
+		{
+			Dispose(true);
+			GC.SuppressFinalize(this);
+		}
+
+		protected virtual void Dispose(bool disposing) { }
 	}
 }